Overlays
Implemented an Actor Faction overlay and a Passable Tiles overlay. I like overlays and will be making a lot of them. Also fixed outdated comments from a part of the codebase that used to be a Roguelike/Racing Game hybrid (a project which did not get far but which I may revisit some day).