A* Implemented
I had written a pretty decent A* implementation for Composelike and wanted to use it here. I simplified it a little, as this project isn't that far along yet, and plugged it in. In many ways this project is the spiritual successor to Composelike. Eventually I want this to work on both platforms. At this point the project has almost reached parity with Composelike, which will be a milestone.