Rebindable Keys
Implemented a hotkey-based method of rebinding keys. User presses CTRL + <Key to Change> and then SHIFT + <new Key>. I chose this solution because it keeps the control scheme keyboard-oriented and does not require any extra menus to implement. Also fixed some outdated comments from much earlier phases of the project (such as a reference to a "Deep Space" tile from when I had an idea that it would take place in space; which may return later when I can do things like add an overworld). As the project evolves I will try to keep the comments accurate but there are probably always going to be a few of those.